refactoring EOCAccordionWithModal out of assessment page for reuse
